Split transformer
